feature/likes

+++ server/redisControllers/exhibition/withdrawalLike.js
@@ -1,26 +1,47 @@
 const { likes } = require("../../models");
 const { Op } = require("sequelize");
 const { isAuthorized } = require("../../utils/tokenFunction");
+const {
+  getCached,
+  caching,
+  delCache,
+} = require("../../utils/redis/cache.ctrl");
 module.exports.withdrawalLike = async (req, res) => {
-  const userInfo = isAuthorized(req);
+  //const userInfo = isAuthorized(req);
+  const userInfo = { id: 45 };
   if (userInfo) {
     const { postId: exhibition_id, type: exhibit_type } = req.params;
     const { id: user_id } = userInfo;
+    const redisKey =
+      exhibit_type === undefined
+        ? "allExhibition"
+        : exhibit_type === "1"
+        ? "standard"
+        : "premium";
+    const reply = await getCached(redisKey);
 
-    /**
-     * type 별로 getCaching
-     * id = exhibition_id 인 데이터 찾아서
-     *
-     */
-    const result = await likes.destroy({
-      where: {
-        [Op.and]: [{ exhibition_id }, { user_id }],
-      },
-    });
+    if (reply) {
+      reply.some((el) => {
+        if (el.id === exhibition_id) {
+          el.likes.some((ele, idx) => {
+            if (ele.user_id === user_id) {
+              el.likes.splice(idx, 1);
+            }
+            return false;
+          });
 
-    if (result) {
+          return false;
+        }
+      });
+
+      caching(redisKey, reply);
       res.status(200).json({
         message: "successfully delete like",
+      });
+      await likes.destroy({
+        where: {
+          [Op.and]: [{ exhibition_id }, { user_id }],
+        },
       });
     }
   } else {
